// snl-gms-mock/packages/weavess-core/src/ts/types/types.ts
export enum DistanceUnits {
  /** String representation of 'degrees' */
  Degrees = 'degrees',

  /** String representation of 'km' */
  Km = 'km'
}
// ...
export interface Station {
  /** Id of station */
  id: string;

  /** Name of station */
  name: string;

  /** Default channel information for station.  Allows for multiple channels for individual mode */
  defaultChannels: Channel[];

  /** Non-default channels for station */
  nonDefaultChannels?: Channel[];

  /** Non-default channels for station */
  splitChannels?: Channel[];

  /** Indicates if the child weavess channels are displayed (label expanded) */
  areChannelsShowing?: boolean;

  /** Distance of station */
  distance?: number;

  /** Units for distance */
  distanceUnits?: DistanceUnits;

  /** Distance of station */
  azimuth?: number;

  /** Has QC masks */
  hasQcMasks?: boolean;
}

/** Specifies a range with a min and max */
export interface Range {
  /** the min */
  min?: number;
  /** the max */
  max?: number;
}

/**
 * An interface describing a channel, including whether it is in an error state.
 * The channel can also take a simple string as a description.
 */
export interface ChannelDescription {
  /** Shown in the bottom right corner of the channel row */
  message: string | undefined;

  /** shown in a tooltip on hover */
  tooltipMessage?: string;

  /** indicates that the description should be formatted as an error state */
  isError?: boolean;
}

/**
 * Channel configuration
 */
export interface Channel {
  /** Id of channel */
  id: string;

  /** Whether this channel is selected */
  isSelected: boolean;

  /** Name of channel */
  name: string;

  /** Azimuth */
  azimuth?: number;

  /** The arrival time of the station nearest the open event */
  baseStationTime?: number;

  /** Label for channels */
  channelLabel?: string;

  /** Optional icon for the channel label */
  channelLabelIcon?: JSX.Element;

  /** Tooltip for the Channel Label */
  channelLabelTooltip?: string;

  /** Tooltip for the Channel Label */
  channelLabelAdditionalInfo?: string;

  /** Type of channel */
  channelType?: string;

  configuredInputChannelNames?: string[];

  /** default display range scale */
  defaultRange?: Range;

  /** The description to show on the bottom right corner of the channel row */
  description?: string | ChannelDescription;

  /** Distance */
  distance?: number;

  /** Distance units */
  distanceUnits?: DistanceUnits;

  /** Height of the channel */
  height?: number;

  /** Optional text to be displayed in the top left of the channel's label */
  labelHeader?: string | JSX.Element;

  /** Collection of markers to be rendered on the channel */
  markers?: Markers;

  /** Spectrogram content */
  spectrogram?: ChannelSpectrogramContent;

  /** The phase of signal detection to be created by the split mode */
  splitChannelPhase?: string;

  /** The split point in the timeline, if split mode is activated */
  splitChannelTime?: number;

  /** The number of seconds the data should be offset  */
  timeOffsetSeconds?: number;

  /** Waveform content */
  waveform?: ChannelWaveformContent;

  /** the y axis ticks */
  yAxisTicks?: number[];
}

export interface ChannelWaveformContent {
  /** Id of channel segment */
  channelSegmentId: string;

  /** Collection of channel segments, keyed on filter name */
  channelSegmentsRecord: Record<string, ChannelSegment[]>;

  /** Collection of markers */
  markers?: Markers;

  /** Collection of masks */
  masks?: Mask[];

  /** Collection of signal detections */
  signalDetections?: PickMarker[];

  /** Collection of predictive phases */
  predictedPhases?: PickMarker[];

  /** Collection of theoretical phase windows */
  theoreticalPhaseWindows?: TheoreticalPhaseWindow[];

  /** Forces the channel to dim */
  isAutoDimmed?: boolean;
}
// ...
export interface ChannelSpectrogramContent {
  /** Spectrogram description */
  description?: string;

  /** Color of the label for the description */
  descriptionLabelColor?: string;

  /** Epoch start time in seconds */
  startTimeSecs: number;

  /** The time step of the spectrogram data (x-axis) */
  timeStep: number;

  /** The frequency step of the spectrogram data (y-axis) */
  frequencyStep: number;

  /**
   * The spectrogram data (time x frequency)
   * Provides the powers or intensity of the spectrogram
   */
  data: number[][];

  /** Collection of markers */
  markers?: Markers;

  /** Collection of signal detections */
  signalDetections?: PickMarker[];

  /** Collection of predictive phases */
  predictedPhases?: PickMarker[];

  /** Collection of theoretical phase windows */
  theoreticalPhaseWindows?: TheoreticalPhaseWindow[];
}
// ...
export interface Markers {
  /** Vertical markers */
  verticalMarkers?: Marker[];

  /** Moveable markers */
  moveableMarkers?: Marker[];

  /** Selection windows */
  selectionWindows?: SelectionWindow[];
}